refactor: SelectJdbcTemplate을 JdbcTemplate으로 통합
